Preserve decimal prices in eBay listings

+++ market_monitor/models/listing.py
@@ -15,7 +15,7 @@ from dataclasses import dataclass
 @dataclass
 class Listing:
     title: str
-    price: int # Decimal
+    price: float
     # currency: str
     url: str
     source: str
